Migrated all signoff components to functional components. Added rjsf types.
