+++ src/components/signoff/Signed.tsx
@@ -1,0 +1,75 @@
+import type { SignoffSourceInfo, DestinationInfo } from "../../types";
+
+import React from "react";
+import AdminLink from "../AdminLink";
+import { ProgressStep } from "./ProgressBar";
+import HumanDate from "./HumanDate";
+
+type SignedProps = {
+  label: string;
+  currentStep: number;
+  step: number;
+  isCurrentUrl: boolean;
+  source: SignoffSourceInfo;
+  destination: DestinationInfo | null;
+};
+
+export function Signed(props: SignedProps) {
+  const { label, currentStep, step, isCurrentUrl, source, destination } = props;
+  return (
+    <ProgressStep label={label} currentStep={currentStep} step={step}>
+      {destination && source.lastSignatureBy && (
+        <SignedInfos
+          source={source}
+          destination={destination}
+          isCurrentUrl={isCurrentUrl}
+        />
+      )}
+    </ProgressStep>
+  );
+}
+
+type SignedInfosProps = {
+  isCurrentUrl: boolean;
+  source: SignoffSourceInfo;
+  destination: DestinationInfo;
+};
+
+function SignedInfos(props: SignedInfosProps) {
+  const { isCurrentUrl, source, destination } = props;
+  const { lastReviewBy, lastReviewDate, lastSignatureBy, lastSignatureDate } =
+    source;
+  const { bid, cid } = destination;
+  return (
+    <ul>
+      <li>
+        <strong>Approved: </strong>
+        <HumanDate timestamp={lastReviewDate} />
+      </li>
+      <li>
+        <strong>By: </strong>
+        {lastReviewBy}
+      </li>
+      {lastReviewDate != lastSignatureDate && (
+        <li>
+          <strong>Re-signed: </strong>
+          <HumanDate timestamp={lastSignatureDate} />
+        </li>
+      )}
+      {lastReviewDate != lastSignatureDate && ( // just to avoid SyntaxError: Adjacent JSX elements
+        <li>
+          <strong>By: </strong>
+          {lastSignatureBy}
+        </li>
+      )}
+      {!isCurrentUrl && (
+        <li>
+          <strong>Destination: </strong>
+          <AdminLink name="collection:records" params={{ bid, cid }}>
+            {`${bid}/${cid}`}
+          </AdminLink>
+        </li>
+      )}
+    </ul>
+  );
+}
